// src/constants/enums.ts
export enum Directions {
  RTL = "rtl",
  LTR = "ltr",
}

export enum Languages {
  ENGLISH = "en",
  ARABIC = "ar",
}

export enum Themes {
  LIGHT = "light",
  DARK = "dark",
}

export enum Routes {
  ROOT = "/",
  AUTH = "auth",
  ADMIN = "admin",
}

export enum Pages {
  SIGNIN = "signin",
  BOOKS = "books",
  STAFF = "staff",

  HERO = "hero",
  ABOUT = "about",
  EVENTS = "events",
  FAQ = "faq",
  EDIT = "edit",
  NEW = "new",
}

export enum InputTypes {
  TEXT = "text",
  EMAIL = "email",
  PASSWORD = "password",
  NUMBER = "number",
  DATE = "date",
  TIME = "time",
  DATE_TIME_LOCAL = "datetime-local",
  CHECKBOX = "checkbox",
  RADIO = "radio",
  SELECT = "select",
  TEXTAREA = "textarea",
  FILE = "file",
  IMAGE = "image",
  COLOR = "color",
  RANGE = "range",
  TEL = "tel",
  URL = "url",
  SEARCH = "search",
  MONTH = "month",
  WEEK = "week",
  HIDDEN = "hidden",
  MULTI_SELECT = "multi select",
}

export enum Environments {
  PROD = "production",
  DEV = "development",
}

export enum UserRole {
  ADMIN = "ADMIN",
  LEADER = "LEADER",
}